#8 Execution with spark
